requirements.txt
scipy~=1.14.1
numpy~=1.26.4
matplotlib~=3.9.2
scikit-learn~=1.5.2
torchvision~=0.19.1
setuptools~=75.1.0
